pyproject.toml
[project]
name = "py-resume"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cyclopts>=4.2.5",
    "jinja2>=3.1.6",
    "playwright>=1.56.0",
    "pydantic>=2.12.4",
    "jinja-markdown>=1.210911",
    "pytest>=9.0.1",
    "pyyaml>=6.0.3",
    "ruff>=0.14.6",
    "ty>=0.0.1a27",
    "pre-commit>=4.5.0",
    "just>=0.8.165",
    "pydantic-ai>=1.22.0",
    "python-dotenv>=1.0.1",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I"]
per-file-ignores = {"resume_generator/assets.py" = ["E501"]}

